refactor cli and plugins to reduce orchestration complexity
